add demo button for myanmar by strategy
