--- app/src/laos/java/org/eyeseetea/malariacare/layout/customization/LoginActivityCustomization.java
@@ -1,9 +1,0 @@
-package org.eyeseetea.malariacare.layout.customization;
-
-import org.eyeseetea.malariacare.LoginActivity;
-
-public class LoginActivityCustomization {
-
-    public void onCreate(LoginActivity loginActivity){
-    }
-}
